erchef: remove most traces of rabbitmq
We simply removed as many references to rabbitmq as we could from the
erchef code base.

We haven't updated all of the tests as they will require some work to
make sure we are still testing some of the creation.

This almost surely will fail at runtime, but it compiles.

Signed-off-by: Steven Danna <steve@chef.io>